dejando lindo los archivos

+++ ejercicio2/source/main.cpp
@@ -3,12 +3,12 @@
 
 
 // comando para compilar
-// g++ abs_combate.cpp abs_guerrero.cpp abs_items.cpp abs_mago.cpp interfaz_personajes.cpp interfaz_armas.cpp main.cpp  -o programa //
+// g++ abs_combate.cpp abs_guerrero.cpp abs_items.cpp abs_mago.cpp interfaz_personajes.cpp interfaz_armas.cpp main.cpp  -o programa 
 
-/*
+
 int main()
 {
-    // menu para corroborar el funcionamiento de los personajes // 
+    // menu para corroborar el funcionamiento de los personajes. contiene valores arbitrarios a modo de ejemplicar el funcionamiento// 
     
     unique_ptr<hechicero> nuevo_hechicero = make_unique<hechicero>(Corte_giratorio, 7, 2, 8);
 
@@ -18,7 +18,7 @@ int main()
     nuevo_hechicero ->modificar_hp(16);
     cout << "vida: " << nuevo_hechicero ->retornar_hp()<< endl; 
 
-    nuevo_hechicero ->agregar_arma(arma_baston, 6, 3, 5.4);
+    nuevo_hechicero ->agregar_arma(arma_baston, 6, 3, static_cast<float>(5.4));
     nuevo_hechicero ->info_arma();
 
     cout << "------------------" << endl;
@@ -31,7 +31,7 @@ int main()
     nuevo_conjurador -> modificar_hp(16);
     cout << "vida: " << nuevo_conjurador ->retornar_hp()<< endl; 
 
-    nuevo_conjurador ->agregar_arma(arma_pocion, 6, 3, 5.4);
+    nuevo_conjurador ->agregar_arma(arma_pocion, 6, 3, static_cast<float>(5.4));
     nuevo_conjurador ->info_arma();
 
     cout << "------------------" << endl;
@@ -44,9 +44,8 @@ int main()
     nuevo_barbaro -> modificar_hp(18);
     cout << "vida: " << nuevo_barbaro ->retornar_hp()<< endl; 
 
-    nuevo_barbaro ->agregar_arma(arma_pocion, 6, 3, 5.4);
+    nuevo_barbaro ->agregar_arma(arma_pocion, 6, 3, static_cast<float>(5.4));
     nuevo_barbaro ->info_arma();
     
    return 0;
 }
-*/
